Added deleting functional

+++ WebAPI/WebAPI/Controllers/AuthorController.cs
@@ -103,8 +103,13 @@ namespace WebAPI.Controllers
         public JsonResult DeleteAuthor(long Id)
         {
             var RequiredAuthor = _db.Authors.SingleOrDefault(A => A.Id == Id);
-            _db.Authors.Remove(RequiredAuthor);
-            _db.SaveChanges();
+
+            if (RequiredAuthor != null)
+            {
+                _db.Authors.Remove(RequiredAuthor);
+                _db.SaveChanges();
+            }
+            
             return new JsonResult(RequiredAuthor);
         }
